school/requirements.txt
-f /usr/share/pip-wheels
Django==2.0.3
djangorestframework==3.7.7
pytz==2018.3
